# lib/api-zod/src/generated/types/calendarRsvpInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ace Digital OS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CalendarRsvpInputStatus } from './calendarRsvpInputStatus';

export interface CalendarRsvpInput {
  status: CalendarRsvpInputStatus;
}